pyproject.toml
[tool.poetry]
name = "osr-stuff"
version = "0.1.0"
description = ""
authors = ["Andrew Herbig <notandrewherbig@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
pygame = "^1.9.6"
recordclass = "^0.13.2"
Pillow = "^7.2.0"

[tool.poetry.dev-dependencies]
black = "^19.10b0"
flake8 = "^3.8.3"
pre-commit = "^2.7.1"
pyupgrade = "^2.7.2"
isort = "^5.4.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
